requirements.txt
pyahocorasick>=2.0.0
pyyaml>=6.0
requests>=2.31.0
anthropic>=0.40.0
python-dateutil>=2.8.0
click>=8.1.0
